+++ frontend/src/widgets/bottom-nav/index.ts
@@ -1,1 +1,2 @@
 export { BottomNavBar } from './ui/BottomNavBar';
+export { NAV_ITEMS, type NavItem } from './nav-items';
