+++ frontend/src/widgets/bottom-nav/ui/BottomNavBar.tsx
@@ -1,21 +1,6 @@
 import { Link, useRouterState } from '@tanstack/react-router';
-import { Home, Search, Music2, User } from 'lucide-react';
 import { cn } from '@/shared/lib/utils';
-import type { ComponentType } from 'react';
-
-export type NavItem = {
-  key: string;
-  label: string;
-  to: string;
-  icon: ComponentType<{ className?: string; 'aria-hidden'?: boolean }>;
-};
-
-export const NAV_ITEMS: NavItem[] = [
-  { key: 'home', label: '홈', to: '/', icon: Home },
-  { key: 'search', label: '검색', to: '/search', icon: Search },
-  { key: 'my-bands', label: '내 밴드', to: '/my-bands', icon: Music2 },
-  { key: 'my', label: '마이', to: '/profile', icon: User },
-];
+import { NAV_ITEMS, type NavItem } from '../nav-items';
 
 type BottomNavBarProps = {
   items?: NavItem[];
